requirements.txt
# Core packages
Flask==2.0.1
pandas==1.5.0
cx-Oracle==8.3.0
python-dotenv==0.19.0
gunicorn==20.1.0

# EXACT versions from model training
scikit-learn==1.5.1
joblib==1.2.0
numpy==1.23.5
scipy==1.9.3
threadpoolctl==3.1.0